Add a description of how to adopt the API
Ref discussion from 20 Aug. The thing I left out of this text
is a global dispatch. The last point made on that was:

    ... there is some inherent difficulties. Local dispatch is the more
    straightforward approach, but difficulties arise when you want a
    particular function to use a particular backend implementation.

and I wasn't quite sure if that concern is valid - maybe I'm missing
some details. If the "you" in there is the library author, then
they should just ignore the whole API standard namespace and just do
`import lib_I_want; lib_I_want.<somefunc>`.
If it is the end user, then they normally would pass in an array of
the right type that would control this. So the one case I see where
the concern applies is if there is some function that does _not_ take
any array input, but creates arrays internally _and_ the default array
type for that creation is somehow undesirable. That seems a bit
unlikely.