Add data interchange section, using DLPack
This is incomplete, but there seems to be agreement that DLPack
is the best candidate for an interchange mechanism. So let's
document that, and improve the content as we go.

See https://github.com/data-apis/consortium-feedback/issues/1
for the RFC about standardizing on DLPack. Design discussion
should be picked up there.