Remove `abs` from table near top of array object specification
`abs` isn't an operator, and it was the only non-operator
left in the top-level table in this file.

Related:

`len(array)` is bad form, and it's not clear what is desired here.
Raising an exception may make sense, but that's not what some
current libraries do.

`abs(array)` also isn't all that great, but seems implemented
all around and is well-defined.