Make axis keyword to squeeze() positional
As suggested by @shoyer in https://github.com/data-apis/array-api/issues/97#issuecomment-747116655
This makes it possible to predict resulting rank of output array, which is
otherwise undetermined (see discussion in gh-97).

Using squeeze without specifying the axis in library code often results in
unintuitive behaviour. For the common use case of extracting a scalar from a size-1 2-D array,
this gets a little more verbose (e.g. `np.squeeze(np.array([[1]]), axis=(0, 1))`),
but that's probably a price worth paying for the extra clarity.

Also changes specified behaviour for a given axis not having size 1 to raising
a `ValueError`, which is what NumPy does. This wasn't considered before, and the current
description seems simply incorrect.

Finally, this makes `squeeze` the exact inverse of `expand_dims`, which is probably
a good thing.