Add contributor acknowledgements via all-contributors
The idea here is to add this on the main repos for arrays and for
dataframes. The initial set of contributors was added via the
all-contributors CLI, later it can be done via the bot.
See https://allcontributors.org/ for details.

It looks like it is not possible at this time to add people who do not
have a GitHub username - I opened an issue on the all-contributors repo
about that.