Added explicit formulae for the factorisations Added properties of the returned matrices in the main description and the return section
Uniform: Always say "(or a stack of matrices)" rather than repeating the
properties of the matrices. E.g. Do not write "a square matrix (or a stack of square
matrices)" but simply "a square matrix (or stack of matrices)"

Uniform: Prefer "Returns" over "Computes" at the start of every
description.

Uniform: Prefer "decomposition" over "factorization"