Require __pow__ and __truediv__ to have floating-point inputs
Previously it was any numeric dtype, but this was inconsistent with the
functions pow() and divide(). Integer inputs are problematic because they
would either always have to be cast to float or use value-based casting
depending on, e.g., the sign of the exponent for __pow__. Furthermore, in
order to properly correspond to Python conventions, __truediv__ should always
return a float (integer division should be implemented by __floordiv__).

This was discussed in #173.