Specify input dtypes for the statistical functions
Functions like mean() that perform divisions require floating-point inputs.
All other functions require numeric inputs.

This was mentioned in #232 but does not fix the main question of that issue
(the behavior of the statistical functions on empty arrays). I can implement
that as well if we come to an agreement about what it should say.