Make the writing of the factorisations in linalg more uniform (#220)
* Added explicit formulae for the factorisations
Added properties of the returned matrices in the main description and
the return section

Uniform: Always say "(or a stack of matrices)" rather than repeating the
properties of the matrices. E.g. Do not write "a square matrix (or a stack of square
matrices)" but simply "a square matrix (or stack of matrices)"

Uniform: Prefer "Returns" over "Computes" at the start of every
description.

Uniform: Prefer "decomposition" over "factorization"

* Remove backticks from equations

The presentation of equations will be fixed in a follow-up PR including TeX rendering support.

* Add qualifier

* Update copy

* Remove line break

* Add qualifiers

Co-authored-by: Athan <kgryte@gmail.com>