Make the Sphinx build error on warnings and fix all warnings (#390)
* Enable nitpicky mode in the Sphinx build

This will make Sphinx warn when there are cross-references that it cannot find
Without it, it just silently makes the text not link to anything.

* Make warnings errors in the Sphinx build

There are several warnings that need to be fixed, but they are all legitimate
errors in the markup that should be fixed.

* Remove duplicated text

* Fix all Sphinx errors from nitpicky mode

A lot of errors come from autodoc wanting to make cross references for every
return type hint. This is disabled manually using the nitpick_ignore variable
in conf.py for each instance.