Describe what extensions are and how to use themo
Note that an alternative idea discussed earlier was a separate function
`get_extension()`, but IIRC that was not considered a good idea.
Now that we have module-level `__getattr__`'s, it should not be
a problem for any library to use a specified name like `linalg` or
`fft`.

Whether the functions in the `linalg` extensions should all be required
to exist or not was discussed in gh-403 (no clear conclusion there).
One option discussed there to deal with hard to implement or more niche
APIs is to create a separate status/label or some other way to track
desired signatures (details to be worked out if we want to go that way).