Fix optional keyword argument in `take` signature (#644)
This commit fixes the function signature for `take`. Namely, when
an input array is one-dimensional, the `axis` kwarg is optional;
when the array has more than one dimension, the `axis` kwarg is
required. Unfortunately, the type signature cannot encode this
duality, and we must rely on the specification text to clarify
that the `axis` kwarg is required for arrays having ranks greater
than unity.

Ref: https://github.com/data-apis/array-api-compat/issues/34