fix: clarify flattening behavior and fix required shape of `inverse_indices`
This commit backports the following changes to the 2021 and 2022
revisions of the Array API standard:

- correction to the required output shape of `inverse_indices`.
Prior revisions stated that this should have the same shape as `x`,
but this is only true if `x` is one-dimensional. For multi-dimensional
`x`, the output shape should the size of a flattened `x`.
- clarifications regarding the contents of returned indices (first
occurrences). Namely, that the returned array should include indices
into a flattened `x`. For the case where `x` is one-dimensional,
the guidance is the same. For multi-dimensional `x`, this clarifies
that returned indices are for the flattened `x` and not the original
input array.