Add recommendation on what to do with DLPack read-only flag (#749)
This is a follow-up to the discussion in gh-191. It's a recommendation
rather than a hard requirement to allow implementers some choice.
That said, this is how things worked in practice before DLPack 1.0
as well, since there was no flag to represent read-only. Experience
with JAX showed that exposing shared memory was preferred by users
over raising or always making a copy of the data on the producer side.